Designer CRUD is running now

+++ app/Services/DesignerService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Exceptions\DesignerHasProjectException;
 use App\Models\Designer;
 use App\Utilities\FileUploadManager;
 use App\Utilities\StorageManager;
@@ -19,23 +20,35 @@ class DesignerService extends CrudService {
     }
 
     public function create($data) {
-        $data['document'] = $this->uploadFile('storage/designers', $data['document']);
+        $data['license'] = $this->uploadFile('storage/designers', $data['license']);
         $this->model->fill($data);
         $this->model->save();
     }
 
-    public function update($data, $model) {
-        if (isset($data['document'])) {
-            $data['document'] = $this->uploadFile($this->path, $data['document']);
-            $this->deleteFile($this->folder, $model->document);
+    /**
+     * @param array $data
+     * @param Designer $model
+     */
+    public function update(array $data, $model) {
+        if (isset($data['license'])) {
+            $data['license'] = $this->uploadFile($this->path, $data['license']);
+            $this->deleteFile($this->folder, $model->license);
         }
 
         $model->fill($data);
         $model->save();
     }
 
+    /**
+     * @param Designer $model
+     * @throws DesignerHasProjectException
+     */
     public function delete($model) {
-        $this->deleteFile($this->folder, $model->document);
+        if ($model->projects->count()) {
+            throw new DesignerHasProjectException();
+        }
+
+        $this->deleteFile($this->folder, $model->license);
         $model->delete();
     }
 }
